courses endpoints now contain authentication

+++ api/courses.js
@@ -3,15 +3,16 @@ const { ValidationError } = require("sequelize");
 
 const router = Router();
 const json2csv = require("json2csv").parse;
-const courses = require('../data/courses');
+const courses = require("../data/courses");
 
 const { Course, CourseClientFields } = require("../models/course");
 const { User } = require("../models/user");
 const { Assignment } = require("../models/assignment");
+const { generateAuthToken, requireAuthentication } = require("../lib/auth");
 
 // Fetch a list of all courses
 router.get("/", async function (req, res, next) {
-  console.log("  -- req.query:", req.query)
+  console.log("  -- req.query:", req.query);
   let page = parseInt(req.query.page) || 1;
   const pageSize = 10;
   const lastPage = Math.ceil(courses.length / pageSize);
@@ -32,7 +33,7 @@ router.get("/", async function (req, res, next) {
   }
   if (page > 1) {
     links.prevPage = `/courses?page=${page - 1}`;
-    links.firstPage = '/courses?page=1';
+    links.firstPage = "/courses?page=1";
   }
 
   /*
@@ -47,22 +48,31 @@ router.get("/", async function (req, res, next) {
       pageSize: pageSize,
       lastPage: lastPage,
       total: courses.length,
-      links: links
+      links: links,
     });
   } catch (e) {
     next(e);
   }
-
 });
 
 // Create a new Course
-// ※※※※ Currently, this endpoint does not contain authentication
 /*******************************************************************
  * Only an authenticated User with 'admin' role can create a new Course.
  ********************************************************************/
-router.post("/", async function (req, res, next) {
+router.post("/", requireAuthentication, async function (req, res, next) {
   try {
+    const user = await User.findByPk(req.user);
+    const isAdmin = user.role === "admin" ? true : false;
+
+    if (!isAdmin) {
+      res
+        .status(403)
+        .send({ error: "Unauthorized to access the specified resource." });
+      return;
+    }
+
     const course = await Course.create(req.body, CourseClientFields);
+
     res.status(201).send({ id: course.id });
   } catch (e) {
     if (e instanceof ValidationError) {
@@ -90,14 +100,13 @@ router.get("/:id", async function (req, res, next) {
 });
 
 // Update data for a specific Course
-// ※※※※ Currently, this endpoint does not contain authentication
 /*******************************************************************
  * Performs a partial update on the data for the Course.
  * Note that enrolled students and assignments cannot be modified via this endpoint.
  * Only an authenticated User with 'admin' role or an authenticated 'instructor' User
  * whose ID matches the instructorId of the Course can update Course information.
  ********************************************************************/
-router.patch("/:id", async function (req, res, next) {
+router.patch("/:id", requireAuthentication, async function (req, res, next) {
   const courseId = req.params.id;
   try {
     const courseData = await Course.findOne({ where: { id: courseId } });
@@ -105,9 +114,26 @@ router.patch("/:id", async function (req, res, next) {
     // if the data does not exist
     if (!courseData) {
       res.status(404).send({ error: "Requested resource does not exist" });
+      return;
     }
 
     try {
+      const user = await User.findByPk(req.user);
+      const isAdmin = user.role === "admin" ? true : false;
+      const isInstructor = user.role === "instructor" ? true : false;
+
+      if (
+        !(
+          isAdmin ||
+          (isInstructor && courseData.dataValues.instructorId === user.id)
+        )
+      ) {
+        res
+          .status(403)
+          .send({ error: "Unauthorized to access the specified resource." });
+        return;
+      }
+
       const result = await Course.update(req.body, {
         where: { id: courseId },
         fields: CourseClientFields,
@@ -130,11 +156,10 @@ router.patch("/:id", async function (req, res, next) {
 });
 
 // Remove a specific Course from the database
-// ※※※※ Currently, this endpoint does not contain authentication
 /*******************************************************************
  * Only an authenticated User with 'admin' role can remove a Course.
  ********************************************************************/
-router.delete("/:id", async function (req, res, next) {
+router.delete("/:id", requireAuthentication, async function (req, res, next) {
   const courseId = req.params.id;
   try {
     const courseData = await Course.findOne({ where: { id: courseId } });
@@ -145,6 +170,16 @@ router.delete("/:id", async function (req, res, next) {
     }
 
     try {
+      const user = await User.findByPk(req.user);
+      const isAdmin = user.role === "admin" ? true : false;
+
+      if (!isAdmin) {
+        res
+          .status(403)
+          .send({ error: "Unauthorized to access the specified resource." });
+        return;
+      }
+
       const result = await Course.destroy({
         where: { id: courseId },
       });
@@ -166,7 +201,6 @@ router.delete("/:id", async function (req, res, next) {
 });
 
 // Fetch a list of the students enrolled in the Course
-// ※※※※ Currently, this endpoint does not contain authentication
 /*******************************************************************
  * Only an authenticated User with 'admin' role or
  * an authenticated 'instructor' User whose ID matches
@@ -186,6 +220,22 @@ router.get("/:id/students", async function (req, res, next) {
     }
   } catch (e) {
     next(e);
+  }
+
+  const user = await User.findByPk(req.user);
+  const isAdmin = user.role === "admin" ? true : false;
+  const isInstructor = user.role === "instructor" ? true : false;
+
+  if (
+    !(
+      isAdmin ||
+      (isInstructor && courseData.dataValues.instructorId === user.id)
+    )
+  ) {
+    res
+      .status(403)
+      .send({ error: "Unauthorized to access the specified resource." });
+    return;
   }
 
   try {
@@ -220,7 +270,6 @@ router.get("/:id/students", async function (req, res, next) {
 });
 
 // Update enrollment for a Course
-// ※※※※ Currently, this endpoint does not contain authentication
 /*******************************************************************
  * Only an authenticated User with 'admin' role or
  * an authenticated 'instructor' User whose ID matches the instructorId
@@ -241,6 +290,22 @@ router.post("/:id/students", async function (req, res, next) {
     const courseData = await Course.findOne({ where: { id: courseId } });
     if (!courseData) {
       res.status(404).send({ error: "Requested resource does not exist" });
+      return;
+    }
+
+    const user = await User.findByPk(req.user);
+    const isAdmin = user.role === "admin" ? true : false;
+    const isInstructor = user.role === "instructor" ? true : false;
+
+    if (
+      !(
+        isAdmin ||
+        (isInstructor && courseData.dataValues.instructorId === user.id)
+      )
+    ) {
+      res
+        .status(403)
+        .send({ error: "Unauthorized to access the specified resource." });
       return;
     }
 
@@ -285,8 +350,6 @@ router.post("/:id/students", async function (req, res, next) {
           removeFailedArray.push(studentId);
         }
       });
-
-      removeFailedArray.push(studentId);
     }
 
     res.status(201).send();
@@ -296,7 +359,6 @@ router.post("/:id/students", async function (req, res, next) {
 });
 
 // Fetch a CSV file containing list of the student enrolled in the Course
-// ※※※※ Currently, this endpoint does not contain authentication
 /*******************************************************************
  * Only an authenticated User with 'admin' role or
  * an authenticated 'instructor' User whose ID matches the instructorId
@@ -315,6 +377,22 @@ router.get("/:id/roster", async function (req, res, next) {
     }
   } catch (e) {
     next();
+  }
+
+  const user = await User.findByPk(req.user);
+  const isAdmin = user.role === "admin" ? true : false;
+  const isInstructor = user.role === "instructor" ? true : false;
+
+  if (
+    !(
+      isAdmin ||
+      (isInstructor && courseData.dataValues.instructorId === user.id)
+    )
+  ) {
+    res
+      .status(403)
+      .send({ error: "Unauthorized to access the specified resource." });
+    return;
   }
 
   try {
@@ -354,7 +432,6 @@ router.get("/:id/roster", async function (req, res, next) {
 });
 
 // Fetch a list of the Assignments for the Course
-// ※※※※ Currently, this endpoint does not contain authentication
 /*******************************************************************
  * Only an authenticated User with 'admin' role or
  * an authenticated 'instructor' User whose ID matches the instructorId
@@ -372,6 +449,22 @@ router.get("/:id/assignments", async function (req, res, next) {
 
   if (!courseData) {
     res.status(404).send({ error: "Requested resource does not exist" });
+    return;
+  }
+
+  const user = await User.findByPk(req.user);
+  const isAdmin = user.role === "admin" ? true : false;
+  const isInstructor = user.role === "instructor" ? true : false;
+
+  if (
+    !(
+      isAdmin ||
+      (isInstructor && courseData.dataValues.instructorId === user.id)
+    )
+  ) {
+    res
+      .status(403)
+      .send({ error: "Unauthorized to access the specified resource." });
     return;
   }
